Use uintptr_t casts rather than int on pointer arithmetic.
Fixes "warning: cast from pointer to integer of diffrent size" on LP64.

XXX: probably we should have HAVE_STDINT_H in config files?

+++ src/malloc.c
@@ -75,6 +75,7 @@ what you give them.   Help stamp out software-hoarding!  */
 /* This must precede sys/signal.h on certain machines.  */
 #include <sys/types.h>
 #include <signal.h>
+#include <stdint.h>
 #else
 
 /* Determine which kind of system this is.  */
@@ -336,8 +337,8 @@ morecore (nu)			/* ask system for more memory */
 	break;
       }
 
-  if ((int) cp & 0x3ff)	/* land on 1K boundaries */
-    sbrk (1024 - ((int) cp & 0x3ff));
+  if ((uintptr_t) cp & 0x3ff)	/* land on 1K boundaries */
+    sbrk (1024 - ((uintptr_t) cp & 0x3ff));
 #endif /* not VMS */
 
  /* Take at least 2k, and figure out how many blocks of the desired size
@@ -359,9 +360,9 @@ morecore (nu)			/* ask system for more memory */
   malloc_sbrk_unused = lim_data - siz;
 
 #ifndef VMS
-  if ((int) cp & 7)
+  if ((uintptr_t) cp & 7)
     {		/* shouldn't happen, but just in case */
-      cp = (char *) (((int) cp + 8) & ~7);
+      cp = (char *) (((uintptr_t) cp + 8) & ~7);
       nblks--;
     }
 #endif /* not VMS */
@@ -392,8 +393,8 @@ getpool ()
   register int nu;
   register char *cp = sbrk (0);
 
-  if ((int) cp & 0x3ff)	/* land on 1K boundaries */
-    sbrk (1024 - ((int) cp & 0x3ff));
+  if ((uintptr_t) cp & 0x3ff)	/* land on 1K boundaries */
+    sbrk (1024 - ((uintptr_t) cp & 0x3ff));
 
   /* Record address of start of space allocated by malloc.  */
   if (_malloc_base == 0)
@@ -658,10 +659,10 @@ memalign (alignment, size)
   if (ptr == 0)
     return 0;
   /* If entire block has the desired alignment, just accept it.  */
-  if (((int) ptr & (alignment - 1)) == 0)
+  if (((uintptr_t) ptr & (alignment - 1)) == 0)
     return ptr;
   /* Otherwise, get address of byte in the block that has that alignment.  */
-  aligned = (char *) (((int) ptr + alignment - 1) & -alignment);
+  aligned = (char *) (((uintptr_t) ptr + alignment - 1) & -alignment);
 
   /* Store a suitable indication of how to free the block,
      so that free can find the true beginning of it.  */
